+++ source/extensions/filters/common/rbac/matchers.cc
@@ -19,7 +19,8 @@ MatcherConstSharedPtr Matcher::create(const envoy::config::rbac::v3::Permission&
   case envoy::config::rbac::v3::Permission::RuleCase::kHeader:
     return std::make_shared<const HeaderMatcher>(permission.header());
   case envoy::config::rbac::v3::Permission::RuleCase::kDestinationIp:
-    return std::make_shared<const IPMatcher>(permission.destination_ip(), true);
+    return std::make_shared<const IPMatcher>(permission.destination_ip(),
+                                             IPMatcher::Type::ConnectionLocal);
   case envoy::config::rbac::v3::Permission::RuleCase::kDestinationPort:
     return std::make_shared<const PortMatcher>(permission.destination_port());
   case envoy::config::rbac::v3::Permission::RuleCase::kAny:
@@ -46,7 +47,11 @@ MatcherConstSharedPtr Matcher::create(const envoy::config::rbac::v3::Principal& 
   case envoy::config::rbac::v3::Principal::IdentifierCase::kAuthenticated:
     return std::make_shared<const AuthenticatedMatcher>(principal.authenticated());
   case envoy::config::rbac::v3::Principal::IdentifierCase::kSourceIp:
-    return std::make_shared<const IPMatcher>(principal.source_ip(), false);
+    return std::make_shared<const IPMatcher>(principal.source_ip(),
+                                             IPMatcher::Type::ConnectionRemote);
+  case envoy::config::rbac::v3::Principal::IdentifierCase::kRemoteIp:
+    return std::make_shared<const IPMatcher>(principal.remote_ip(),
+                                             IPMatcher::Type::DownstreamRemote);
   case envoy::config::rbac::v3::Principal::IdentifierCase::kHeader:
     return std::make_shared<const HeaderMatcher>(principal.header());
   case envoy::config::rbac::v3::Principal::IdentifierCase::kAny:
@@ -123,10 +128,21 @@ bool HeaderMatcher::matches(const Network::Connection&,
 }
 
 bool IPMatcher::matches(const Network::Connection& connection, const Envoy::Http::RequestHeaderMap&,
-                        const StreamInfo::StreamInfo&) const {
-  const Envoy::Network::Address::InstanceConstSharedPtr& ip =
-      destination_ ? connection.localAddress() : connection.remoteAddress();
-
+                        const StreamInfo::StreamInfo& info) const {
+  Envoy::Network::Address::InstanceConstSharedPtr ip;
+  switch (type_) {
+  case ConnectionLocal:
+    ip = connection.localAddress();
+    break;
+  case ConnectionRemote:
+    ip = connection.remoteAddress();
+    break;
+  case DownstreamRemote:
+    ip = info.downstreamRemoteAddress();
+    break;
+  default:
+    NOT_REACHED_GCOVR_EXCL_LINE;
+  }
   return range_.isInRange(*ip.get());
 }